# swagger.yaml
consumes:
    - application/json
produces:
    - application/json
schemes:
    - https
swagger: "2.0"
info:
    contact:
        email: test@trahan.dev
        name: Justin Trahan
    license:
        name: N/A
    termsOfService: there are no TOS at this moment, use at your own risk we take no responsibility
    title: go-infra API.
    version: v1.1.0
host: infra.trahan.dev
paths:
...
    /secrets/create:
        post:
            operationId: createUserSecret
            parameters:
                - in: body
                  name: body
                  schema:
                    $ref: '#/definitions/CreateSecretRequest'
                  x-go-name: Body
            responses:
                "200":
                    $ref: '#/responses/CreateSecretResponse'
                "400":
                    description: Invalid request
                "401":
                    description: Unauthorized
            summary: Create a new external application secret.
            tags:
                - secrets
    /secrets/delete/{ID}:
        delete:
            operationId: deleteUserSecretByID
            parameters:
                - in: path
                  name: ID
                  required: true
                  type: string
            responses:
                "200":
                    description: Secret deleted successfully
                "401":
                    description: Unauthorized
                "403":
                    description: Forbidden
                "404":
                    description: Not Found
            summary: Delete a user secret by ID.
            tags:
                - secrets
    /secrets/{ID}:
        get:
            operationId: getUserSecretByID
            parameters:
                - description: ID of secret
                  in: path
                  name: ID
                  required: true
                  type: string
            responses:
                "200":
                    $ref: '#/responses/RetrievedSecretResponse'
                "401":
                    description: Unauthorized
                "403":
                    description: Forbidden
                "404":
                    description: Not Found
            summary: Retrieve a user secret by ID.
            tags:
                - secrets
...
    /user/secrets/by-name/{APPNAME}/{USERID}:
        get:
            operationId: GetUserSecretEntriesByAppName
            parameters:
                - format: uuid
                  in: path
                  name: USERID
                  required: true
                  type: string
                - in: path
                  name: APPNAME
                  required: true
                  type: string
            responses:
                "200":
                    $ref: '#/responses/GetUserSecretEntriesResponse'
                "400":
                    description: Invalid request
                "401":
                    description: Unauthorized
                "403":
                    description: Forbidden
                "404":
                    description: Not Found
                "500":
                    description: Internal Server Error
            summary: Retrieve a user secret by USERID and application name.
            tags:
                - secrets
    /user/secrets/{USERID}:
        get:
            operationId: GetUserSecretEntries
            parameters:
                - format: uuid
                  in: path
                  name: USERID
                  required: true
                  type: string
            responses:
                "200":
                    $ref: '#/responses/GetUserSecretEntriesResponse'
                "401":
                    description: Unauthorized
                "403":
                    description: Forbidden
                "404":
                    description: Not Found
            summary: Retrieve a user secret by USERID.
            tags:
                - secrets
    /user/{APPID}/secrets/{USERID}:
        get:
            operationId: GetUserSecretEntriesByAppId
            parameters:
                - format: uuid
                  in: path
                  name: USERID
                  required: true
                  type: string
                - format: uuid
                  in: path
                  name: APPID
                  required: true
                  type: string
            responses:
                "200":
                    $ref: '#/responses/GetUserSecretEntriesResponse'
                "401":
                    description: Unauthorized
                "403":
                    description: Forbidden
                "404":
                    description: Not Found
            summary: Retrieve a user secret by USERID.
            tags:
                - secrets
...
definitions:
...
    CreateSecretRequest:
        properties:
            application_id:
                format: uuid
                type: string
                x-go-name: ApplicationID
            expiration:
                format: date-time
                type: string
                x-go-name: Expiration
            secret:
                type: string
                x-go-name: Secret
        type: object
        x-go-package: github.com/babbage88/go-infra/services/user_secrets
...
    ExternalAppSecretMetadata:
        properties:
            createdAt:
                format: date-time
                type: string
                x-go-name: CreatedAt
            expiry:
                format: date-time
                type: string
                x-go-name: Expiration
            id:
                format: uuid
                type: string
                x-go-name: Id
            userId:
                format: uuid
                type: string
                x-go-name: UserId
        type: object
        x-go-package: github.com/babbage88/go-infra/services/user_secrets
...
    ExternalApplicationInfo:
        properties:
            id:
                format: uuid
                type: string
                x-go-name: Id
            name:
                type: string
                x-go-name: Name
            url:
                type: string
                x-go-name: UrlEndpoint
        type: object
        x-go-package: github.com/babbage88/go-infra/services/user_secrets
...
    UserSecretEntry:
        properties:
            appInfo:
                $ref: '#/definitions/ExternalApplicationInfo'
            secretMetadata:
                $ref: '#/definitions/ExternalAppSecretMetadata'
        type: object
        x-go-package: github.com/babbage88/go-infra/services/user_secrets
responses:
...
    CreateSecretResponse:
        description: (empty)
        schema:
            properties:
                id:
                    format: uuid
                    type: string
                    x-go-name: ID
            type: object
...
    GetUserSecretEntriesResponse:
        description: (empty)
        schema:
            items:
                $ref: '#/definitions/UserSecretEntry'
            type: array
...
    RetrievedSecretResponse:
        description: (empty)
        schema:
            properties:
                expiration:
                    format: date-time
                    type: string
                    x-go-name: Expiration
                external_application_id:
                    format: uuid
                    type: string
                    x-go-name: ExternalApplication
                id:
                    format: uuid
                    type: string
                    x-go-name: ID
                secret:
                    type: string
                    x-go-name: Secret
                user_id:
                    format: uuid
                    type: string
                    x-go-name: UserID
            type: object
...
securityDefinitions:
    bearer:
        in: header
        name: Authorization
        type: apiKey
security:
    - bearer: []